requirements.txt
django
channels
redis
channels-redis
djangorestframework
paramiko
python-gssapi
Pillow
django-crispy-forms
django-otp
#MySQL-python==1.2.5
django-cors-headers
djangorestframework-simplejwt
pycrypto
python-magic
cryptography
qrcode
supervisor